# server-stub/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: support@onevoice.com
    name: OneVoice Team
  description: AI-powered video dubbing service backend API
  license:
    name: MIT
  title: OneVoice Backend API
  version: 1.0.0
servers:
- description: Production server
  url: https://api.onevoice.com
- description: Staging server
  url: https://staging-api.onevoice.com
- description: Development server
  url: http://localhost:8000
security:
- bearerAuth: []
...
paths:
...
  /api/videos:
    get:
      parameters:
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: per_page
        required: false
        schema:
          default: 20
          type: integer
        style: form
      - explode: true
        in: query
        name: status
        required: false
        schema:
          enum:
          - uploaded
          - processing
          - completed
          - failed
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_api_videos_get_200_response"
          description: Videos retrieved
      summary: Get user videos
      tags:
      - Video Management
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_api_videos_post_request"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
          description: Video created successfully
      summary: Create new video
      tags:
      - Video Management
  /api/videos/{id}:
    delete:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_api_auth_forgot_password_post_200_response"
          description: Video deleted successfully
      summary: Delete video
      tags:
      - Video Management
    get:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
          description: Video retrieved
      summary: Get video by ID
      tags:
      - Video Management
    put:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_api_videos__id__put_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
          description: Video updated successfully
      summary: Update video
      tags:
      - Video Management
  /api/videos/upload-url:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_api_videos_upload_url_post_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_api_videos_upload_url_post_200_response"
          description: Upload URL generated
      summary: Get pre-signed upload URL
      tags:
      - Video Management
  /api/videos/upload:
    post:
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/_api_videos_upload_post_request"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
          description: Video uploaded successfully
      summary: Upload video file
      tags:
      - Video Management
  /api/videos/{id}/dub:
    post:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DubRequest"
        required: true
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_api_videos__id__dub_post_202_response"
          description: Dubbing process started
      summary: Start video dubbing process
      tags:
      - Video Management
  /api/videos/{id}/status:
    get:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatus"
          description: Processing status retrieved
      summary: Get video processing status
      tags:
      - Video Management
  /api/videos/{id}/cancel:
    post:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_api_auth_forgot_password_post_200_response"
          description: Processing cancelled
      summary: Cancel video processing
      tags:
      - Video Management
  /api/videos/{id}/download:
    get:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - explode: true
        in: query
        name: language
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_api_videos__id__download_get_200_response"
          description: Download URL generated
      summary: Download processed video
      tags:
      - Video Management
  /api/videos/{id}/duplicate:
    post:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
          description: Video duplicated successfully
      summary: Duplicate video
      tags:
      - Video Management
  /api/videos/{id}/thumbnail:
    post:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/_api_videos__id__thumbnail_post_request"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_api_videos__id__thumbnail_post_200_response"
          description: Thumbnail generated successfully
      summary: Generate or update video thumbnail
      tags:
      - Video Management
...
  /api/videos/voices:
    get:
      parameters:
      - explode: true
        in: query
        name: language
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Voice"
                type: array
          description: Available voices retrieved
      summary: Get available voices for language
      tags:
      - Video Management
components:
  schemas:
...
    Video:
      example:
        duration: 0.8008281904610115
        original_language: original_language
        target_languages:
        - target_languages
        - target_languages
        video_url: https://openapi-generator.tech
        updated_at: 2000-01-23T04:56:07.000+00:00
        description: description
        created_at: 2000-01-23T04:56:07.000+00:00
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        title: title
        thumbnail_url: https://openapi-generator.tech
        status: uploaded
      properties:
        id:
          format: uuid
          title: id
          type: string
        title:
          title: title
          type: string
        description:
          title: description
          type: string
        duration:
          title: duration
          type: number
        original_language:
          title: original_language
          type: string
        target_languages:
          items:
            type: string
          title: target_languages
          type: array
        status:
          enum:
          - uploaded
          - processing
          - completed
          - failed
          title: status
          type: string
        thumbnail_url:
          format: uri
          title: thumbnail_url
          type: string
        video_url:
          format: uri
          title: video_url
          type: string
        created_at:
          format: date-time
          title: created_at
          type: string
        updated_at:
          format: date-time
          title: updated_at
          type: string
      title: Video
      type: object
    DubRequest:
      example:
        voice_id: voice_id
        target_language: target_language
        preserve_background_music: true
      properties:
        target_language:
          title: target_language
          type: string
        voice_id:
          title: voice_id
          type: string
        preserve_background_music:
          default: true
          title: preserve_background_music
          type: boolean
      required:
      - target_language
      - voice_id
      title: DubRequest
      type: object
    JobStatus:
      example:
        error_message: error_message
        estimated_completion: 2000-01-23T04:56:07.000+00:00
        job_id: job_id
        progress: 8.008281904610115
        status: pending
      properties:
        job_id:
          title: job_id
          type: string
        status:
          enum:
          - pending
          - processing
          - completed
          - failed
          - cancelled
          title: status
          type: string
        progress:
          maximum: 100
          minimum: 0
          title: progress
          type: number
        estimated_completion:
          format: date-time
          title: estimated_completion
          type: string
        error_message:
          title: error_message
          type: string
      title: JobStatus
      type: object
...
    Voice:
      example:
        gender: male
        preview_url: https://openapi-generator.tech
        name: name
        language: language
        style: style
        id: id
      properties:
        id:
          title: id
          type: string
        name:
          title: name
          type: string
        language:
          title: language
          type: string
        gender:
          enum:
          - male
          - female
          - neutral
          title: gender
          type: string
        style:
          title: style
          type: string
        preview_url:
          format: uri
          title: preview_url
          type: string
      title: Voice
      type: object
...
    _api_auth_forgot_password_post_200_response:
      example:
        message: message
      properties:
        message:
          title: message
          type: string
      title: _api_auth_forgot_password_post_200_response
      type: object
...
    _api_videos_get_200_response:
      example:
        pagination: "{}"
        videos:
        - duration: 0.8008281904610115
          original_language: original_language
          target_languages:
          - target_languages
          - target_languages
          video_url: https://openapi-generator.tech
          updated_at: 2000-01-23T04:56:07.000+00:00
          description: description
          created_at: 2000-01-23T04:56:07.000+00:00
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          title: title
          thumbnail_url: https://openapi-generator.tech
          status: uploaded
        - duration: 0.8008281904610115
          original_language: original_language
          target_languages:
          - target_languages
          - target_languages
          video_url: https://openapi-generator.tech
          updated_at: 2000-01-23T04:56:07.000+00:00
          description: description
          created_at: 2000-01-23T04:56:07.000+00:00
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          title: title
          thumbnail_url: https://openapi-generator.tech
          status: uploaded
      properties:
        videos:
          items:
            $ref: "#/components/schemas/Video"
          title: videos
          type: array
        pagination:
          title: pagination
          type: object
      title: _api_videos_get_200_response
      type: object
    _api_videos_post_request:
      properties:
        title:
          title: title
          type: string
        description:
          title: description
          type: string
        original_language:
          title: original_language
          type: string
      required:
      - title
      title: _api_videos_post_request
      type: object
    _api_videos__id__put_request:
      properties:
        title:
          title: title
          type: string
        description:
          title: description
          type: string
      title: _api_videos__id__put_request
      type: object
    _api_videos_upload_url_post_request:
      properties:
        filename:
          title: filename
          type: string
        content_type:
          title: content_type
          type: string
      required:
      - content_type
      - filename
      title: _api_videos_upload_url_post_request
      type: object
    _api_videos_upload_url_post_200_response:
      example:
        upload_url: https://openapi-generator.tech
        video_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        upload_url:
          format: uri
          title: upload_url
          type: string
        video_id:
          format: uuid
          title: video_id
          type: string
      title: _api_videos_upload_url_post_200_response
      type: object
    _api_videos_upload_post_request:
      properties:
        file:
          format: binary
          type: string
        title:
          type: string
        description:
          type: string
      type: object
    _api_videos__id__dub_post_202_response:
      example:
        job_id: job_id
        message: message
      properties:
        job_id:
          title: job_id
          type: string
        message:
          title: message
          type: string
      title: _api_videos__id__dub_post_202_response
      type: object
    _api_videos__id__download_get_200_response:
      example:
        expires_at: 2000-01-23T04:56:07.000+00:00
        download_url: https://openapi-generator.tech
      properties:
        download_url:
          format: uri
          title: download_url
          type: string
        expires_at:
          format: date-time
          title: expires_at
          type: string
      title: _api_videos__id__download_get_200_response
      type: object
    _api_videos__id__thumbnail_post_request:
      properties:
        thumbnail:
          format: binary
          type: string
        timestamp:
          description: Time in seconds to extract thumbnail from video
          type: number
      type: object
    _api_videos__id__thumbnail_post_200_response:
      example:
        thumbnail_url: https://openapi-generator.tech
      properties:
        thumbnail_url:
          format: uri
          title: thumbnail_url
          type: string
      title: _api_videos__id__thumbnail_post_200_response
      type: object
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http